test/mochitest/test_blobfree_bad.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1233029
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  Components.utils.import("resource://gre/modules/FileUtils.jsm");
  Components.utils.import("resource://gre/modules/osfile.jsm");

  let invalidZipRoot = FileUtils.getDir("TmpD", ["b2g-installer-test"], true).path;
  let invalidZip = OS.Path.join(invalidZipRoot, "blobfreeInvalidZip.zip");

  function generate_invalid_zip() {
    ok(true, "generating broken zip file")
    let invalid = new FileUtils.File(invalidZip);
    if (invalid.exists()) {
      ok(false, "There is already a file " + invalidZip);
      return Promise.reject();
    }

    ok(true, "return writeAtomic()")
    return OS.File.writeAtomic(invalidZip, "IAMNOTABLOBFREE");
  }

  function remove_invalid_zip() {
    ok(true, "removing broken zip file")
    let invalid = new FileUtils.File(invalidZip);
    if (!invalid.exists()) {
      ok(false, "There is no file " + invalidZip);
      return Promise.reject();
    }

    ok(true, "return removeDir()")
    return OS.File.removeDir(invalidZipRoot);
  }

  function test_extract_failure() {
    ok(true, "in promise with " + invalidZip);
    return extractBlobFreeDistribution(invalidZip)
      .then((root) => {
        ok(false, "in promise .then() with " + invalidZip);
        // if we reach .then() then it's not good
        Promise.reject();
      })
      .catch((err) => {
        ok(true, "in promise .catch() with " + invalidZip);
        Promise.resolve();
      });
  }

  function runTest() {
    loadAboutJs()
      .then(() => generate_invalid_zip())
      .then(() => test_extract_failure())
      .then(() => remove_invalid_zip())
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
